use xorshift to create an infinite scroller
create a grid of towers that scrolls endlessly past the camera. Use
xorshift to create a unique random sequence for each grid location that
controls the height and rotation of the tower. Wrap the glowing coloured
cubes to within fixed x,z bounds of the camera i.e as they disappear
behind the camera they re-appear in front.

TODO
* LRU caching of tower parameters to avoid re-evaluating xorshift
random chains.
* changing direction of scroll (smooth turns?, collision avoidance?)
* user control